// packages/datadog-ci/src/commands/synthetics/interfaces.ts
export interface DatadogCIConfig extends APIHelperConfig {
  configPath: string
}

// eslint-disable-next-line @typescript-eslint/no-empty-interface
export interface SyntheticsCIConfig extends DatadogCIConfig {}
// ...
export interface ImportTestsCommandConfig extends SyntheticsCIConfig {
  configPath: string
  files: string[]
  publicIds: string[]
  testSearchQuery?: string
}